remove submodule
