+++ src/rgw/rgw_common.h
@@ -2326,97 +2326,5 @@ struct obj_time {
 };
 
 
-class RGWLatency {
-  private:
-    static struct obj_time client_to_rgw_time[1000];
-    static int client_to_rgw_count;
-    static struct obj_time rgw_to_client_time[1000];
-    static int rgw_to_client_count;
-    static struct obj_time rgw_to_rados_time[1000];
-    static int rgw_to_rados_count;
-    static struct obj_time rados_to_rgw_time[1000];
-    static int rados_to_rgw_count;
-    
-    static std::mutex file_write_mutex;
-
-  public:
-    static void get_time_client_to_rgw(const std::string& name, const struct timespec ts) {
-      client_to_rgw_time[client_to_rgw_count].used = 1;
-      client_to_rgw_time[client_to_rgw_count].name = name;
-      client_to_rgw_time[client_to_rgw_count].tv_sec = ts.tv_sec;
-      client_to_rgw_time[client_to_rgw_count].tv_nsec = ts.tv_nsec;
-      client_to_rgw_count++;
-      
-    }
-    static void get_time_rgw_to_client(const std::string& name, const struct timespec ts) {
-      rgw_to_client_time[rgw_to_client_count].used = 1;
-      rgw_to_client_time[rgw_to_client_count].name = name;
-      rgw_to_client_time[rgw_to_client_count].tv_sec = ts.tv_sec;
-      rgw_to_client_time[rgw_to_client_count].tv_nsec = ts.tv_nsec;
-      rgw_to_client_count++;
-    }
-    static void get_time_rgw_to_rados(const std::string& name, const struct timespec ts) {
-      rgw_to_rados_time[rgw_to_rados_count].used = 1;
-      rgw_to_rados_time[rgw_to_rados_count].name = name;
-      rgw_to_rados_time[rgw_to_rados_count].tv_sec = ts.tv_sec;
-      rgw_to_rados_time[rgw_to_rados_count].tv_nsec = ts.tv_nsec;
-      rgw_to_rados_count++;
-    }
-    static void get_time_rados_to_rgw(const std::string& name, const struct timespec ts) {
-      rados_to_rgw_time[rados_to_rgw_count].used = 1;
-      rados_to_rgw_time[rados_to_rgw_count].name = name;
-      rados_to_rgw_time[rados_to_rgw_count].tv_sec = ts.tv_sec;
-      rados_to_rgw_time[rados_to_rgw_count].tv_nsec = ts.tv_nsec;
-      rados_to_rgw_count++;
-    }
-
-    static void init_times() {
-      memset(&client_to_rgw_time, 0, sizeof(struct obj_time)*1000);
-      memset(&rgw_to_client_time, 0, sizeof(struct obj_time)*1000);
-      memset(&rgw_to_rados_time, 0, sizeof(struct obj_time)*1000);
-      memset(&rados_to_rgw_time, 0, sizeof(struct obj_time)*1000);
-      client_to_rgw_count = 0;
-      rgw_to_client_count = 0;
-      rgw_to_rados_count = 0;
-      rados_to_rgw_count = 0;
-    }
-
-    static uint64_t to_nsec(const time_t sec, const long nsec) {
-      return (uint64_t)nsec + (uint64_t)sec * 1000000000ull;
-    }
-
-    static void print_file(struct obj_time* time, std::string path) {
-      std::ofstream f;
-      f.open(path, std::ios_base::app);
-      if (f.is_open())  {
-        for (int i = 0; i < 1000; i++) {
-          if ((time+i)->used != 1)
-            break;
-          auto name = (time+i)->name;
-          //auto sec = std::to_string((time+i)->tv_sec).c_str();
-          //auto nsec = std::to_string((time+i)->tv_nsec).c_str();
-          auto nsec = to_nsec( (time+i)->tv_sec, (time+i)->tv_nsec );
-          f << name;
-          //f.write(name, sizeof(name));
-          f << '\t';
-          f << nsec;
-          //f.write(nsec, sizeof(nsec));
-          f << '\n';
-        }
-      }
-      f.close();
-    }
-  
-    static int time_file_dump() {
-      std::scoped_lock lock(file_write_mutex);
-      print_file(client_to_rgw_time, "/tmp/client_to_rgw_time.txt");
-      print_file(rgw_to_client_time, "/tmp/rgw_to_client_time.txt");
-      print_file(rgw_to_rados_time, "/tmp/rgw_to_rados_time.txt");
-      print_file(rados_to_rgw_time, "/tmp/rados_to_rgw_time.txt");
-      init_times();
-    return 0; 
-    }
-};
-
 
 #endif
